add `quat_dist` and fix `quat_err`
I had very strange relation with `quat_err`. I had replicated a result of attitude control system from Markley and Crassisis' book using the old implementation. I had done "a fix" with "my intuition" and implemented different expression for error from what was suggested in the book. It turns out that the fix, with an added change—the flip of arguments, did the trick. Minus and minus resulted in plus.