fix error on shepperd’s algorithm
A very humbling experience to find a fault and realize that a very basic function had remained untested. At the same time, it is also a reminder of how faulty this codebase could be — and a reminder to work harder/smarter/better ...

I tested and generalized Shepperd's algorithm to reduce multiple branching. However, [1] suggests that Markley's improvement makes the algorithm more robust to the deviation of the input matrix from orthogonal matrix:

"Shepperd's algorithm is guaranteed to produce a precisely normalized quaternion only if A is precisely orthogonal. This modification of Shepperd's algorithm also provides a very efficient method for computing an exactly orthogonal matrix that is close to an approximately orthogonal matrix."

Then why fix the bug while if you plan to implement other algorithm? I intend this commit to be future reference should I ever require implementation of Shepperd's algorithm or desire to revert.

References:
  [1] Markley - Unit quaternion from rotation matrix (2008)