add function to restore the orthogonality of almost dcm
Markley's paper on DCM to quaternion has a section called "Orthogonalization," where he suggests orthogonalizing a matrix by converting the DCM to a quaternion and back to DCM — that's exactly what we do.

He also makes an important point about the unnecessary normalization inside `dcm_to_quat()` if its output is going to be converted back to a DCM. It would make things more efficient, but I'm not sure on how I would handle it because I definitely do not want to pass "normalize_flag" into `dcm_to_quat()`.