fix the issue on `quat_to_dcm()` and revelation on `dcm_to_quat()`
`dcm_to_quat()` with _noisy_ input DCM goes crazy. I experienced it first hand trying to replicate the results of [1]. [2] makes it evident that, for non orthogonal matrix, we have to compute quaternion using the approaches similar to the solutions of Wahba's problem---and it makes sense to me. However, I am unable to replicate the results of [1] using [3] which is supposed to be robust to non orthogonal input DCM.

[1] Mao - Optimal orthonormalization of the strapdown matrix by using singular value decomposition (1986)
[2] Bar-Itzhack - New Method for Extracting the Quaternion from a Rotation Matrix (2000)
[3] Markley - Unit quaternion from rotation matrix (2008)